GPKG: detect float->int32 overflow in ComputeTileAndPixelShifts()
git-svn-id: https://svn.osgeo.org/gdal/trunk@37537 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/gdrivers/gpkg.py
@@ -1367,6 +1367,20 @@ def gpkg_14():
         return 'fail'
     ds = None
 
+    # Overflow occured in ComputeTileAndPixelShifts()
+    with gdaltest.error_handler():
+        ds = gdal.OpenEx('tmp/tmp.gpkg', open_options = ['MINX=-1e12', 'MAXX=-0.9999e12'])
+    if ds is not None:
+        gdaltest.post_reason('fail')
+        return 'fail'
+
+    # Overflow occured in ComputeTileAndPixelShifts()
+    with gdaltest.error_handler():
+        ds = gdal.OpenEx('tmp/tmp.gpkg', open_options = ['MINY=-1e12', 'MAXY=-0.9999e12'])
+    if ds is not None:
+        gdaltest.post_reason('fail')
+        return 'fail'
+
     os.remove('tmp/tmp.gpkg')
     return 'success'
